Extend response entity id from int to long

+++ src/main/java/cz/silesnet/sis/sync/item/reader/ResponseId.java
@@ -5,14 +5,14 @@ public class ResponseId {
   private final String textId;
   private final String timeStamp;
   private final String sequenceNo;
-  private final int id;
+  private final long id;
 
   private ResponseId(String textId) {
     this.textId = textId;
     String[] fields = textId.split("_");
     timeStamp = fields[0];
     sequenceNo = fields[1];
-    id = Integer.valueOf(fields[2]);
+    id = Long.valueOf(fields[2]);
   }
 
   public static ResponseId of(String id) {
@@ -27,7 +27,7 @@ public class ResponseId {
     return sequenceNo;
   }
 
-  public int id() {
+  public long id() {
     return id;
   }
